Changed from Double Wrapper to BigDecimal Wrapper

+++ src/main/java/coj/and/CaloriesCalculator/aliments/Aliments.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -29,23 +30,23 @@ public class Aliments {
     private String name;
     @Column(name = "calories", nullable = false, columnDefinition = "REAL")
     @Min(value = 0, message = "Calories should not be negative")
-    private Double calories;
+    private BigDecimal calories;
     @Column(name = "protein", nullable = false, columnDefinition = "REAL")
     @Min(value = 0, message = "Protein should not be negative")
-    private Double protein;
+    private BigDecimal protein;
     @Column(name = "carbs", nullable = false, columnDefinition = "REAL")
     @Min(value = 0, message = "Carbs should not be negative")
-    private Double carbs;
+    private BigDecimal carbs;
     @Column(name = "fat", nullable = false, columnDefinition = "REAL")
     @Min(value = 0, message = "Fat should not be negative")
-    private Double fat;
+    private BigDecimal fat;
     @Column(name = "fiber", nullable = false, columnDefinition = "REAL")
     @Min(value = 0, message = "Fiber should not be negative")
-    private Double fiber;
+    private BigDecimal fiber;
     @OneToMany(mappedBy = "aliments", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<UserAliments> userAliments = new ArrayList<UserAliments>();
 
-    public Aliments(String name, Double calories, Double protein, Double carbs, Double fat, Double fiber) {
+    public Aliments(String name, BigDecimal calories, BigDecimal protein, BigDecimal carbs, BigDecimal fat, BigDecimal fiber) {
         this.name = name;
         this.calories = calories;
         this.protein = protein;
